refactor: post/battle 구조 변경
