+++ src/pages/post/detail.tsx
